restructured httpd's folder structure
