+++ kevin/httpd/__init__.py
@@ -1,0 +1,3 @@
+"""
+Web API for kevin
+"""
